Move block codec registry to the header
